Work on various things, mostly for implementing the employee list.

+++ src/Employee.java
@@ -1,9 +1,8 @@
-
 public class Employee extends User{
 
     public Employee(String surname, String firstname, String gender,
                     String email, String roleName, int empNum,
-                    int phoneNum, String dob, char[] password) {
+                    long phoneNum, String dob, char[] password) {
         super(surname, firstname, gender, email, roleName, empNum, phoneNum, dob, password);
 
     }
