+++ src/HR.java
@@ -1,8 +1,7 @@
-
 public class HR extends User{
 
     public HR(String surname, String firstname, String gender,
-              String email, String roleName, int empNum, int phoneNum, String dob, char[] password) {
+              String email, String roleName, int empNum, long phoneNum, String dob, char[] password) {
         super(surname, firstname, gender, email, roleName, empNum, phoneNum, dob, password);
 
     }
