+++ src/SalaryWorker.java
@@ -1,10 +1,8 @@
-
-
 public class SalaryWorker extends Employee{
     private float yearlySalary;
 
     public SalaryWorker(String surname, String firstname, String gender,
-                        String email, String roleName, int empNum, int phoneNum, String dob,
+                        String email, String roleName, int empNum, long phoneNum, String dob,
                         float salary, char[] password) {
         super(surname, firstname, gender, email, roleName, empNum, phoneNum,
                 dob, password);
