+++ src/WageWorker.java
@@ -1,10 +1,8 @@
-
-
 public class WageWorker extends Employee{
     private float hourlyWage;
 
     public WageWorker(String surname, String firstname, String gender,
-                      String email, String roleName, int empNum, int phoneNum, String dob,
+                      String email, String roleName, int empNum, long phoneNum, String dob,
                       float wage, char[] password) {
         super(surname, firstname, gender, email, roleName, empNum, phoneNum,
                 dob, password);
